modify to follow the coding standard
