--- src/main/java/FriendlyTool/process/ftException.java
@@ -1,7 +1,0 @@
-package FriendlyTool.process;
-
-public class ftException extends  Exception {
-    public ftException(String message) {
-        super(message);
-    }
-}
