+++ src/main/java/friendlytool/process/FTException.java
@@ -1,0 +1,7 @@
+package friendlytool.process;
+
+public class FTException extends Exception {
+    public FTException(String message) {
+        super(message);
+    }
+}
