+++ src/main/java/friendlytool/process/Storage.java
@@ -1,7 +1,7 @@
-package FriendlyTool.process;
+package friendlytool.process;
 
-import FriendlyTool.command.Parser;
-import FriendlyTool.task.*;
+import friendlytool.command.Parser;
+import friendlytool.task.*;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -9,7 +9,7 @@ import java.io.IOException;
 import java.util.Scanner;
 
 public class Storage {
-    public static void loadTask(TaskList myList) throws ftException {
+    public static void loadTask(TaskList myList) throws FTException {
         try {
             File save = new File("./data/myTask.txt");
             if (!save.exists()) {
@@ -24,26 +24,28 @@ public class Storage {
                 String[] parsedSave = Parser.parseSave(s.nextLine());
                 String taskType = parsedSave[0];
                 switch (taskType) {
-                    case "T":
-                        myList.add(new ToDo(parsedSave[2], Parser.parseBool(parsedSave[1])));
-                        break;
-                    case "D":
-                        myList.add(new Deadline(parsedSave[2], Parser.parseBool(parsedSave[1]), new Date(parsedSave[3])));
-                        break;
-                    case "E":
-                        myList.add(new Event(parsedSave[2], Parser.parseBool(parsedSave[1]), new Date(parsedSave[3]), new Date(parsedSave[4])));
-                        break;
-                    default:
-                        throw new ftException("    Warning: The file is corrupted. Please delete the file");
+                case "T":
+                    myList.add(new ToDo(parsedSave[2], Parser.parseBool(parsedSave[1])));
+                    break;
+                case "D":
+                    myList.add(new Deadline(parsedSave[2], Parser.parseBool(parsedSave[1]),
+                            new Date(parsedSave[3])));
+                    break;
+                case "E":
+                    myList.add(new Event(parsedSave[2], Parser.parseBool(parsedSave[1]), new Date(parsedSave[3]),
+                            new Date(parsedSave[4])));
+                    break;
+                default:
+                    throw new FTException("    Warning: The file is corrupted. Please delete the file");
                 }
                 s.nextLine();
             }
         } catch (IOException e) {
-            throw new ftException("File not found");
+            throw new FTException("File not found");
         }
     }
 
-    public static void updateTask(TaskList list) throws ftException {
+    public static void updateTask(TaskList list) throws FTException {
         try {
             FileWriter fw = new FileWriter("./data/myTask.txt");
             for (int i = 0; i < list.size(); i++) {
@@ -52,7 +54,7 @@ public class Storage {
             }
             fw.close();
         } catch (IOException e) {
-            throw new ftException("Error in updating the task");
+            throw new FTException("Error in updating the task");
         }
     }
