+++ src/main/java/friendlytool/task/ToDo.java
@@ -1,4 +1,4 @@
-package FriendlyTool.task;
+package friendlytool.task;
 
 public class ToDo extends Task {
     public ToDo(String name, boolean isDone) {
@@ -13,9 +13,5 @@ public class ToDo extends Task {
     @Override
     public String toSaveFormat() {
         return "T " + super.toSaveFormat() + "\n";
-    }
-
-    public boolean equals(ToDo t) {
-        return t.getName() == this.getName();
     }
 }
